Add files via upload
This code has two different speed controllers: the same one found in the digital joystick code ("Joe Motor Code") and another one that I made to reduce the sensitivity of the analog joystick. As well, the code includes the same functionality as the original with the incorporation of an analog potentiometer. - Joe